+++ test/api/api_test.cpp
@@ -1489,8 +1489,8 @@ source %s/mod_{24572..28666}
   ASSERT_EQ(1, r);
   ASSERT_EQ(DS_ERROR_KIND_TYPE_ERROR, e->kind);
   ASSERT_STREQ("GlobalLimit", e->name);
-  ASSERT_EQ(9, e->lineNum);
-  ASSERT_EQ(8, e->chars);
+  ASSERT_EQ(1, e->lineNum);
+  ASSERT_EQ(5, e->chars);
   e = newError();
 }